Control.Agent.Free

Description

Agents for free.

Documentation

type Agent f t = FreeT f (t Identity)

An Agent f t a is a program which uses functor f as a low-level API, monad transformer t — for high-level features and which returns a value of type a.

The important thing is that the environment may mess with t on initialization of the agent or on low-level API calls. This was designed for the possibility of behaviosites (see http://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.114.4071 for more details).

type Agent' f = Agent f IdentityT

An Agent without any exposed structure.

execAgent :: (MFunctor t, MonadTrans t, Monad m, Functor f, Monad (t m), Monad (t Identity)) => (forall b. f b -> m b) -> Agent f t a -> t m a

Execute an agent program with particular interpreter. execAgent int agent give an interpretation to the low-level API.

execAgent' :: (Monad m, Functor f) => (forall b. f b -> m b) -> Agent' f a -> m a

Like execAgent for agents without exposed structure.

execAgent_ :: (MFunctor t, MonadTrans t, Monad m, Functor f, Monad (t m), Monad (t Identity)) => (forall b. f b -> m b) -> Agent f t a -> t m ()

Like execAgent but discards the result.

execAgent'_ :: (Monad m, Functor f) => (forall b. f b -> m b) -> Agent' f a -> m ()

Like execAgent' but discards the result.

transform :: (MonadTrans t, Monad m, Functor f, Monad (t m), Monad (t Identity)) => (forall b. f b -> Agent f t b) -> Agent f t a -> Agent f t a

transform phi agent applies phi to each low-level API command in agent program. This is the basis for behaviosites.