Control.Agent.Free.Internal

Description

Internals of the free agent.

Documentation

type Agent t f = Demarcate t (Free f)

An Agent t f a is a program which uses functor f as a low-level API, monad transformer t — for high-level features and which returns a value of type a.

The important thing is that the environment may mess with t on initialization of the agent or on low-level API calls. This was designed for the possibility of behaviosites (see http:citeseerx.ist.psu.eduviewdocsummary?doi=10.1.1.114.4071 for more details).

type Agent' f = Agent IdentityT f

An Agent without any exposed structure.

runAgent :: (Functor f, Monad (t (Free f)), MonadTrans t) => Agent t f a -> t (Free f) a

Turn agent program into a t (Free f) a free monadic value.

liftCmd :: (Functor f, MonadFree f m, MonadTrans t) => f a -> t m a

Turn an API functor into a polymorphic DSL command for agent programming. An example of typical use:

 data API next
    = Output String next
    | Input (String -> next)
    deriving (Functor)

output s = liftCmd $ Output s ()
 input    = liftCmd $ Input id